Move to filepathx package
